+++ src/ops.rs
@@ -11,7 +11,7 @@
  *  You should have received a copy of the GNU General Public License
  *  along with this program.  If not, see <https://www.gnu.org/licenses/>.
  */
-use crate::core::{Args, AmbosoEnv, AmbosoMode, AmbosoLintMode, INVIL_VERSION, INVIL_OS, EXPECTED_AMBOSO_API_LEVEL, parse_stego_toml, SemVerKey};
+use crate::core::{Args, AmbosoEnv, AmbosoMode, AmbosoLintMode, INVIL_VERSION, INVIL_OS, EXPECTED_AMBOSO_API_LEVEL, parse_stego_toml, lex_stego_toml, SemVerKey};
 use crate::utils::try_parse_stego;
 use std::process::Command;
 use std::io::{self, Write};
@@ -667,6 +667,8 @@ pub fn do_query(env: &AmbosoEnv, args: &Args) -> Result<String,String> {
                         let mut kulpotests_map = env.kulpotests_table.clone();
                         alltests_map.append(&mut bonetests_map);
                         alltests_map.append(&mut kulpotests_map);
+                        let mut tot_successes = 0;
+                        let mut tot_failures = 0;
                         for test in alltests_map.values() {
                             if test.exists() {
                                 trace!("Found {{{}}}", test.display());
@@ -680,7 +682,16 @@ pub fn do_query(env: &AmbosoEnv, args: &Args) -> Result<String,String> {
                                             let test_elapsed = env.start_time.elapsed();
                                             info!("Done test {{{}}}, Elapsed: {:.2?}", test.display(), test_elapsed);
                                         }
-                                        info!("Test cmd: {{{:?}}}", test_res);
+                                        match test_res {
+                                            Ok(st) => {
+                                                info!("Test ok: {st}");
+                                                tot_successes += 1;
+                                            }
+                                            Err(e) => {
+                                                error!("Test {} failed. Err: {e}", test.display());
+                                                tot_failures += 1;
+                                            }
+                                        }
                                     } else {
                                         debug!("{} is not executable", test.display());
                                         return Ok("Is not executable".to_string());
@@ -694,8 +705,14 @@ pub fn do_query(env: &AmbosoEnv, args: &Args) -> Result<String,String> {
                                 return Err("No file found".to_string())
                             }
                         }
-                        info!("Done test macro");
-                        return Ok("Done test macro run".to_string());
+                        debug!("Done test macro");
+                        info!("Successes: {tot_successes}");
+                        error!("Failures: {tot_failures}");
+                        if tot_failures != 0 {
+                            return Err("Test macro had some failures".to_string());
+                        } else {
+                            return Ok("Done test macro run".to_string());
+                        }
                     }
                 }
                 _ => {
@@ -763,7 +780,7 @@ pub fn run_test(test_path: &PathBuf, record: bool) -> Result<String,String> {
                                     }
                                 }
                             } else {
-                                info!("Expected: {{\"\n{}\"}}", stdout_record);
+                                warn!("Expected: {{\"\n{}\"}}", stdout_record);
                                 match std::str::from_utf8(&output.stdout) {
                                     Ok(v) => {
                                         info!("Found: {{\"\n{}\"}}", v);
@@ -799,7 +816,7 @@ pub fn run_test(test_path: &PathBuf, record: bool) -> Result<String,String> {
                             info!("Stderr matched!");
                         } else {
                             warn!("Stderr did not match!");
-                            info!("Expected: {{\"\n{}\"}}", stderr_record);
+                            warn!("Expected: {{\"\n{}\"}}", stderr_record);
                             match std::str::from_utf8(&output.stderr) {
                                 Ok(v) => {
                                     info!("Found: {{\"\n{}\"}}", v);
@@ -1012,6 +1029,19 @@ pub fn handle_linter_flag(stego_path: &PathBuf, lint_mode: AmbosoLintMode) -> Re
                     }
                     Err(e) => {
                         error!("Failed lint for {{{}}}.\nError was:    {e}",stego_path.display());
+                        return Err(e);
+                    }
+                }
+            }
+            AmbosoLintMode::Lex => {
+                let res = lex_stego_toml(stego_path);
+                match res {
+                    Ok(_) => {
+                        info!("Lex successful for {{{}}}.", stego_path.display());
+                        return Ok("Linter lex check success".to_string());
+                    }
+                    Err(e) => {
+                        error!("Failed lex for {{{}}}.\nError was:    {e}",stego_path.display());
                         return Err(e);
                     }
                 }